refactor(graphic): 优化DX12图形适配器实现
- 将DX12纹理宽高类型从uint32_t改为FLOAT以提高精度
- 更新纹理子资源尺寸计算使用uint64_t防止溢出
- 修改着色器转换使用dynamic_cast增强类型安全
- 在管道状态设置中改进空指针检查逻辑
- 移除冗余头文件引用并重新组织包含顺序
- 统一BOOL类型定义兼容Windows和非Windows平台
- 为纹理描述结构体字段更新为uint64_t类型
- 修正命令缓冲区中视口高度设置的数据类型
- 为ITexture接口添加nodiscard属性提升可读性
- 调整混合因子枚举值删除无效的BlendFactor项

+++ src/engine/graphic/adapters/dx12/DX12CommandBuffer.cpp
@@ -137,7 +137,7 @@ void DX12CommandBuffer::BeginRenderPass(const RenderPassDesc& desc) {
         // 设置视口和裁剪矩形
         D3D12_VIEWPORT viewport = {};
         viewport.Width = static_cast<float>(dx12RenderTarget->GetWidth());
-        viewport.Height = static_cast<float>(dx12RenderTarget->GetHeight());
+        viewport.Height = dx12RenderTarget->GetHeight();
         viewport.MinDepth = 0.0f;
         viewport.MaxDepth = 1.0f;
         m_commandList->RSSetViewports(1, &viewport);
